requirements.txt
pyspark==3.5.3
setuptools
requests==2.32.3
delta-spark==3.2.1
pathlib2==2.3.6
pyarrow==12.0.1
pandas==2.0.2
numpy==1.24.2
